services/graph-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Kafka
aiokafka==0.10.0

# Neo4j
neo4j==5.16.0

# Redis
redis==5.0.1

# Google Gemini AI
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
